♻️ (#417) Task 타입 구조 개선
